+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - Associated type `Indexer::Data` for `indexer::BlockEventsIndexer`.
   - Method `Client::get_block_items` response stream items.
   - Method `Client::get_finalized_block_item` return type.
+  - Type `PendingUpdate` field `effect`.
 
 - Add `NextUpdateSequenceNumbers::protocol_level_tokens` and protobuf deserialization of it
 - Changed `TokenClient`'s `burn` and `mint` methods to accept a singular `TokenAmount`, instead of `Vec<TokenAmount>`.